baidunews/requirements.txt
scrapy
newspaper3k
pymongo
